split ssz test
